Rename `zattrs_utils.py` into `utils.py`
