+++ fractal_tasks_core/tasks/yokogawa_to_ome_zarr.py
@@ -32,7 +32,7 @@ from fractal_tasks_core.ngff import load_NgffImageMeta
 from fractal_tasks_core.ome_zarr.channels import get_omero_channel_list
 from fractal_tasks_core.ome_zarr.channels import OmeroChannel
 from fractal_tasks_core.ome_zarr.pyramids import build_pyramid
-from fractal_tasks_core.ome_zarr.zattrs_utils import (
+from fractal_tasks_core.ome_zarr.utils import (
     get_parameters_from_metadata,
 )
 from fractal_tasks_core.roi import check_valid_ROI_indices
